requirements.txt
Django>=3.0.0,<3.1.0
djangorestframework>=3.10.3,<3.11.0